Isochronous transfer mode is no longer user-selectable.
A VideoStreaming interface can support either bulk or isochronous
transfer, but not both. Since the control block specifies a VS interface and
we can tell which method the interface supports, there's no point in making
the user choose the correct mode.

I've replaced the 'isochronous' parameter to uvc_start_streaming() and
uvc_stream_start() with a 'flags' parameter that may be used in the future
to select additional stream behavior. The lower bit is reserved, so users
that still set flags=1, intending to select isochronous mode, will be ignored.