header file
sorry ken, not suer whether including the linuvc_internal header is a
bad idea. it defines the type uvc_streaming_interface_t which I was
hoping to use in returning the list of available frame formats. would
you be so kind to explain whether good idea or not? or if not, how to
get allow use of the type in diag.c