redo pull request
hi ken, sorry i screwed things up in the last request and started over.
i had adding a call to request the list of stream formats but I think
the better strategy is to just go with the routines that get stream
formats for a particular device handle inside of device.c. i could not
figure out a good way to reverse the changes i had made after i had
committed them to my master. the addition of a print formats in drag is
useful, though.