add basic UVC 1.50 support
The size of stream control probe/commit messages changed with UVC 1.50.
In theory, the spec allows the webcam to remain functional even if the
host uses the incorrect size.
Yet, some webcams (such as as the one found in a Thinkpad T490s,
pid:vid=04f2:b67c) just respond with a stall error.