Use an already opened device handle to read descriptors (#201)
When opening the uvc device it's opened again in uvc_get_device_descriptor() function.
This change adds a complementary function that reuses current device handle to read descriptors.

Also, it's required if uvc_wrap() function is used to create the uvc device on Android. In this case, the uvc_get_device_descriptor() will fail.