Opt-out kernel driver detachment on interface claim
On macOS Monterey, kernel driver detachment  is required to get camera access. Only libusb 1.0.25 or newer implements this on macOS. A side effect is that it either requires root privileges or the "com.apple.vm.device-access" entitlement.

On earlier versions of macOS, the kernel detachment is not necessary. Therefore, this commit allows to opt-out of this procedure to avoid the necessity of requiring root privileges.