Reduce default number of transfer bufs for MacOS builds
On isochronous transfers, libusb_submit_transfer will choke when
more than 27 buffers are queued at a time. It has done for every
major MacOS release for years, and continues to be the case.

This change reducees the default number of transfer buffers for Mac
builds to a safe number with some margin.