PREVENT reading memory over allocated buffer
In case of too heavy CPU load on HOST the image could not
be completely transmitted from the camera to the HOST. User callback
function is called anyway and in case conversion from YUYV into BGR
format is called in user callback function then memory can be read over
allocated buffer. Namely there was not check if input frame was complete.